Changed meta/main.yml dependencies to match the role definition in playbook
In the test.yml playbook, roles with tags are defined as

  - { role: a_tagged, tags: a }

but the meta/main.yml file for each role DID NOT match that.

  - { role: a_tagged }

So ansible saw those two as DIFFERENT instantiations of the role, and
ran them multiple times.  Once per the playbook definition (with the
tags: bits) and one per the dependency definition WITHOUT the tags:

So, best way to handle dependencies ...

* Apply tags to the roles in the playbook
* Ensure meta/main.yml dependency role definition MATCH those